moving validator into validators.py
